Move `ErrorBox` and `boxError` into separate file
